requirements.txt
Django==4.0.3
djangorestframework==3.13.1
drf_rw_serializers==1.0.4
huey==2.4.3
sympy==1.10.1
